Change select2 initialization/redraw logic
To avoid mithril redraw causing issues with select2, the select2
initialization function is called on child select elements to reload the
select2, when a parent select element selects an element, or is cleared.

Fix a bug where the loop populating requiredByFilters was called on
each redraw, causing the same filters to appear several times in the
list. Now requiredByFilters and requiredLevels are defined in a new
callback from Collection.get.